requirements.txt
python-dotenv
viam-sdk==0.27.0